# website/static/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectorDetail": {
        "properties": {
          "connector": {
            "$ref": "#/components/schemas/ConnectorHealthRow"
          },
          "recent_runs": {
            "items": {
              "$ref": "#/components/schemas/RunRow"
            },
            "title": "Recent Runs",
            "type": "array"
          },
          "runs_page": {
            "$ref": "#/components/schemas/RunsPage"
          },
          "status_filter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Filter"
          }
        },
        "required": [
          "connector",
          "recent_runs",
          "runs_page"
        ],
        "title": "ConnectorDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpsSnapshot": {
        "properties": {
          "connectors": {
            "items": {
              "$ref": "#/components/schemas/ConnectorHealthRow"
            },
            "title": "Connectors",
            "type": "array"
          },
          "recent_runs": {
            "items": {
              "$ref": "#/components/schemas/RunRow"
            },
            "title": "Recent Runs",
            "type": "array"
          },
          "runs_page": {
            "$ref": "#/components/schemas/RunsPage"
          },
          "summary": {
            "$ref": "#/components/schemas/OpsSummary"
          }
        },
        "required": [
          "summary",
          "connectors",
          "recent_runs",
          "runs_page"
        ],
        "title": "OpsSnapshot",
        "type": "object"
      },
...
      "PushResponse": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "integer"
          },
          "rejected": {
            "title": "Rejected",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "accepted",
          "rejected",
          "run_id"
        ],
        "title": "PushResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "IngestRelay",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ops/connectors/{connector_id}": {
      "get": {
        "operationId": "ops_connector_detail_ops_connectors__connector_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "title": "Connector Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ops Connector Detail"
      }
    },
    "/ops/runs/{run_id}": {
      "get": {
        "operationId": "ops_run_detail_ops_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ops Run Detail"
      }
    },
...
    "/studio/connectors/{connector_id}/clone": {
      "get": {
        "operationId": "studio_connector_clone_studio_connectors__connector_id__clone_get",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "title": "Connector Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Studio Connector Clone"
      }
    },
    "/studio/connectors/{connector_id}/delete": {
      "get": {
        "operationId": "studio_connector_delete_studio_connectors__connector_id__delete_get",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "title": "Connector Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Studio Connector Delete"
      }
    },
    "/studio/connectors/{connector_id}/edit": {
      "get": {
        "operationId": "studio_connector_edit_studio_connectors__connector_id__edit_get",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "title": "Connector Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Studio Connector Edit"
      }
    },
    "/studio/connectors/{connector_id}/pause": {
      "get": {
        "operationId": "studio_connector_pause_studio_connectors__connector_id__pause_get",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "title": "Connector Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Studio Connector Pause"
      }
    },
    "/studio/connectors/{connector_id}/resume": {
      "get": {
        "operationId": "studio_connector_resume_studio_connectors__connector_id__resume_get",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "title": "Connector Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Studio Connector Resume"
      }
    },
    "/studio/connectors/{connector_id}/run": {
      "get": {
        "operationId": "studio_connector_run_studio_connectors__connector_id__run_get",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "title": "Connector Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Studio Connector Run"
      }
    },
    "/v1/connectors/{connector_id}/events": {
      "post": {
        "operationId": "push_events_v1_connectors__connector_id__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "title": "Connector Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Push Events"
      }
    },
    "/v1/ops/connectors/{connector_id}": {
      "get": {
        "operationId": "ops_connector_snapshot_v1_ops_connectors__connector_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "title": "Connector Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit_runs",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit Runs",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset_runs",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset Runs",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ops Connector Snapshot"
      }
    },
    "/v1/ops/runs/{run_id}": {
      "get": {
        "operationId": "ops_run_snapshot_v1_ops_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ops Run Snapshot"
      }
    },
    "/v1/ops/snapshot": {
      "get": {
        "operationId": "ops_snapshot_v1_ops_snapshot_get",
        "parameters": [
          {
            "in": "query",
            "name": "window_hours",
            "required": false,
            "schema": {
              "default": 168,
              "maximum": 720,
              "minimum": 1,
              "title": "Window Hours",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit_runs",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit Runs",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset_runs",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset Runs",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "connector_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Connector Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ops Snapshot"
      }
    },
...
    "/v1/studio/connectors/preview": {
      "post": {
        "operationId": "studio_preview_draft_v1_studio_connectors_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Studio Preview Draft"
      }
    },
...
  }
}